+++ packages/core/src/runtime/test-observers/testTypes.ts
@@ -1,0 +1,2 @@
+export { default as Test } from '../Test'
+export { ITest } from '../../interface/ITest'